Fix layout animations flickering on Android [New Architecture] (#6518)
## Summary

This PR fixes issues with layout animation flickering on Android. The
issue was caused by overwriting the `layoutAnimations_` on the JS
thread, before the layout animation is started on the UI thread. On iOS
this logic is always executed on the UI thread - that's why this issue
only occurred on Android. To combat this I moved all updates to
`layoutAninmations_` to the UI thread.
<table>
<tr><td width="50%">before</td><td width="50%">after</td></tr>
<tr><td>


https://github.com/user-attachments/assets/06c5c0fd-df0e-432a-9c54-2edd5fd6b7c4

</td><td>


https://github.com/user-attachments/assets/dc7de243-597b-48a8-bc65-a3e9d3c352c3

</td></tr>
</table>

## Test plan
Check in BB example if there is a flicker when triggering a layout
animation during an entering animation. Also stress test 2048 example.