Move `ReaLayoutAnimator` to a separate file (#6530)
## Summary

This PR moves class `ReaLayoutAnimator` from
`ReanimatedNativeHierarchyManager.java` to its own file
`ReaLayoutAnimator.java`. That's because the signature of one of the
methods in `ReanimatedNativeHierarchyManager` was changed in RN 0.76 and
we'll need to create a sourceset for this very class. I also added
`@Override` annotation to ensure we're overriding this method.

## Test plan