Add `HostFunction` handling to `_scheduleOnJS` (#6571)
## Summary

Currently, when a Host Function is called through `runOnJS` the app
crashes, since we are only handling Remote Function in the
`_scheduleOnJS` method. For example the following code leads to a crash:

```js
useEffect(() => {
    const id = setInterval(() => {
      console.log('interval');
    }, 1000);
    const intervalAlias = clearInterval;
    runOnUI(() => {
      runOnJS(intervalAlias)(id);
    })();
  }, []);
```

This example works fine on the old architecture, because Timer functions
are polyfilled there. On the New Architecture they are Host Functions,
so we need to handle them separately.

## Test plan

Check

```js
useEffect(() => {
    const id = setInterval(() => {
      console.log('interval');
    }, 1000);
    const intervalAlias = clearInterval;
    runOnUI(() => {
      runOnJS(intervalAlias)(id);
    })();
  }, []);
```