fix: build error with c++17 (#6590)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
There was a small typo in #6553 which caused building with c++17 to
still fail.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Build works again with RN 0.72 which uses c++17.

## Notes

- Fixes
https://github.com/software-mansion/react-native-reanimated/issues/6512#issuecomment-2407118212
- Fixes #6584