Bump reanimated version to 3.16.0 (#6599)
## Summary

This PR bumps the Reanimated version to 3.16.0 and updates the
compatibility table (there is also a minor Keyframe fix). This PR
doesn't follow the usual release PR scheme, as we want to change the way
we do releases. Now the `next` reanimated version will be used on main
and release package will be made from a separate branch.

## Test plan