[LA] Update window update check (#6618)
## Summary

This PR changes the way we keep track of the window dimensions in Layout
Animations on the New Arch. The previous version didn't work on Android
in release mode. This new approach is possible because we moved the
initialization of layout animations to the commit hook in #6586.

## Test plan

Check if Layout Animations dependent on `windowWidth` (e.g. SlideInLeft)
work properly.