feat(Worklets): Create a stub of Worklets Module (#6539)
This pull request is one of many which replace

- #6378

since that PR is extensive and too difficult to review.

## Summary

Adding a second Native Module for Worklets.

It's initial responsibility for now is forwarding `valueUnpackerCode` to
Reanimated Module. While it seems silly, it's a good start since it sets
up the whole pipeline of:

```tree
typescript
├── android
│   └── cpp
└── ios
    └── cpp
```
(I was too lazy to use mermaid, long live `mkdir -p`)

Follow up PRs will move more and more responsibilities from Reanimated
to Worklets.

Requires:

- #6556
- #6557 

## Test plan

- [x] All GitHub Actions pass
- [x] Compatibility Github Action passes
- [x] debug Android works
- [x] release Android works
- [x] debug iOS works
- [x] release iOS works