Fix  `MountingOverrideDelegate` initialization (#6841)
## Summary

Currently we use the `setMountingOverrideDelegate` function to pass
`LayoutAnimationProxy` to every surface, when we find out that a new
surface was created. However, the `setMountingOverrideDelegate` function
actually doesn't override the `MountingOverrideDelegate` - it stores all
of them. This would lead to us processing the same transaction multiple
times, so now we skip each already configured surface.
## Test plan