fix: `useAnimatedProps` on input component not working on web  (#6964)
## Summary

Fixes #6202.
Due to React Native's `_valueTracker`, using `useAnimatedProps` won't
work on input component. Before when using `react-native-web` <= 18 we
were able to set `text` props using `setNativeProps` legacy function.
Currently on newer `react-native-web`, (and due to the `_valueTracker`),
we were trying to `setAttribute` `text` on input component
-unsuccessfully.
This PR implements fix that checks whether current component is `input`
with a prop `text`, if so it updates props directly on it.

With a support from this
[PR](https://github.com/software-mansion/react-native-reanimated/pull/6691).

## Test plan

Run **JS props** or **Amount** in `web-example`

---------

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>