(cherry-pick) fix: boxShadow sometimes not working on web (#7358)
…w sometimes not working for web (#7338)This PR fixes boxShadow
sometimes not working on web. To work boxShadow in the form of string
must have `offsetX`, `offsetY`, `blurRadius` and `spreadDistance`. When
one of these props was missing the boxShadow wasn't being processed ergo
was staying as string - that's why it sometimes **worked** on web. This
PR checks if we are on web and if no we then can process the boxShadow.

Additionally I've replaced `shadow` with `boxShadow` in ColorExample.tsx

Check ColorExample on web