fix: Initial animated style updater style is applied on re-render (#7376)
## Summary

This PR fixes #7375.

### The problem

After removing style flattening the old implementation started being
invalid because the `_initialStyle` object that was created during the
initial render of the component contained props from all animated
styles. If a view had more than just one animated style, this object
contained props from all of these styles.

Let's say we had these 3 animated styles:

```tsx
const aStyle1 = useAnimatedStyle(() => ({})); // empty

const aStyle2 = useAnimatedStyle(() => ({
  height: 100
}));

const aStyle3 = useAnimatedStyle(() => ({
  width: 100
}));

<Animated.View style={[aStyle1, aStyle2, aStyle3]} />
```

The style array on the first render looked like this:

```
[{}, { height: 100 }, { height: 100, width: 100 }] // subsequent objects contain all props from previous objects because of merging
```

On the next render, however, the `_initialStyle` was used for all
animated styles, so we ended up with this array:

```
[{ height: 100, width: 100 }, { height: 100, width: 100 }, { height: 100, width: 100 }]
```

As a result, 2 of 3 styles changed. Because the implementation is
different on Paper and we have no props registry (similar to what we
have on Fabric), this change resulted in the view's style update, even
though the animated style wasn't recalculated and didn't return a new
style. This is an incorrect behavior resulting in problems described in
the attached issue.

### The solution

I decided to create a map with style objects and their initial run
values. As a result, instead of returning the same single style object
with all style props from all animated styles, we can now return proper
initial style values for a specific animated style.

## Test plan

Copy the example from #7375 issue and see that it now works.